refactor v2 eval number ops
